Exercício 108 feito

+++ curso-de-python/mundo-3/video-cursos/curso-python-022/ex107/ex107.py
@@ -1,6 +1,6 @@
 import moeda 
 
-preco = int(input('Digite o preço R$: '))
+preco = float(input('Digite o preço R$: '))
 print(f'A metade {preco} é {moeda.metade(preco)}')
 print(f'O dobro de {preco} é {moeda.dobro(preco)}')
 print(f'Aumentando 10%, temos {moeda.aumentar(preco, 10)}')
